Add files via upload
In this case, I use the skin cancer dataset for the case of image classification using CNN Deep Learning. From this dataset, you are assigned to build a CNN model with the provisions of the Balance, Balance, and Balance Augmentation dataset testing scheme. Explore as much as possible to build your best CNN model. Pay attention to the following scoring points for more details.