Add files via upload
In previous modules, we learned how to build a CNN architecture from scratch. Creating your own architecture is indeed more flexible. We can add and remove layers as we like. However, creating your own architecture is sometimes a hassle. Because we have to set (tune) to so many parameters. There is a better way to create a deep learning model using the CNN method, namely by utilizing the model that has been created by previous researchers. These models are proven to be able to produce good accuracy and error values, and are ranked at the top in several competitions. Some models are already provided by 'Keras' library, and we just have to use them. The process of reusing an existing model, to be installed in a different dataset, is known as Transfer Learning